Tests/LibWeb/Text/expected/selection-drag-into-textless-overlay.txt

selection anchor stayed in text: true
selection focus stayed in text: true
